+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org).
 
+## 0.3.0 - 2017-02-09
+
+### Added
+
+  - Selection sort, tests and implementation.
+  - Insertion sort, tests and implementation.
+
 ## 0.2.0 - 2017-02-08
 
 ### Added